requirements.txt
firebase-admin==6.2.0
gspread==6.2.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
tensorflow==2.16.1
Flask==2.3.3
APScheduler==3.10.4
requests==2.31.0
schedule==1.2.1
